refactor(#446): extract MemoryStoreState/Actions types into memoryStoreTypes.ts
